+++ tests/Fixtures/EnumWidgetService.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Ngos\AdminCore\Tests\Fixtures;
+
+use Ngos\AdminCore\Services\BaseService;
+
+class EnumWidgetService extends BaseService
+{
+    public function __construct(EnumWidget $widget)
+    {
+        $this->model = $widget;
+    }
+}